python/app/asig: Initial commit
The application already presents a fairly complete audio signal generator.
It allows to mix up to four common wave forms as well as a form of noise.
Shape, frequency, and amplitudes of signals and noise are configurable.

Apart from straight audio output, output to a UDP socket as well as raw
file and wave file generation are supported.

The application depends on wxPython, GNU Radio, and psutils.