python/app/alab: Initial commit
Alab is meant as an application to explore the effect of GNU Radio
components when applied to audio frequencies. Currently it only supports
common filters.

Audio input can come from either the Mic In port of the sound card, from a
prerecorded raw file, or from raw data received over a UDP connection.
Incoming as well as processed signals are displayed as scope FFT, and
waterfall views. Optional audio output is supported as well.

There are still some layout problems in particular with respect to the
pane where the filters are displayed. However, I regard the user interface
good enough to be released. Of course, any fixes are always welcome.